requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.2.0
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flatbuffers==24.3.25
gast==0.4.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.1
h5py==3.11.0
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
keras
libclang==18.1.1
markdown==3.6
markupsafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlalchemy==2.0.31
tensorflow==2.10.0
termcolor==2.4.0
urllib3==2.2.2
werkzeug==3.0.3
wrapt==1.16.0
wtforms==3.1.2
zipp==3.19.2
